+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Added support for `#[serde(deny_unknown_fields)]` to all configs.
+- Added support for `#[serde(rename_all_fields)]` to enums.
+
 ## 0.17.3
 
 #### 🐞 Fixes